requirements.txt
langchain>=1.0.0
langchain-community>=0.4.0
langgraph>=1.0.0
langgraph-cli[inmem]>=0.4.0
langchain-openai>=0.4.0
python-dotenv>=1.0.0
requests>=2.32.0
